+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.0 (07.01.17)
+*  Add production build support
+*  Add build and stats script for webpack to generate stats.json 
+
 ## 0.4.0 (27.12.16)
 *  Add Redux support
 *  Add support for thunks through redux-pack